D2-PR: Add the integration tests for sessionservice

+++ sportcenter/src/main/java/ca/mcgill/ecse321/sportcenter/service/SessionService.java
@@ -19,7 +19,6 @@ import ca.mcgill.ecse321.sportcenter.repository.LocationRepository;
 import ca.mcgill.ecse321.sportcenter.repository.SessionRepository;
 import ca.mcgill.ecse321.sportcenter.repository.SportCenterRepository;
 import jakarta.transaction.Transactional;
-import java.util.List;
 
 /*
 * <p>Service class in charge of managing sessions. It implements following use cases: </p>
@@ -135,18 +134,22 @@ public class SessionService {
 
     @Transactional
     //An instructor or an owner cancel a session
-    public void cancelSession(int sid){
+    public boolean cancelSession(int sid){
         Session sessionToCancel = sessionRepo.findById(sid);
+        if(sessionToCancel == null){
+            return false;
+        }
         sessionRepo.delete(sessionToCancel);
         //I am unsure if this line is necessary
         sessionToCancel.delete();
+        return true;
 
     }
 
     //--------------------------// Getters //--------------------------//
 
     @Transactional
-    public List<Session> findAllSessions() {
+    public List<Session> findAllSessions() { 
         return toList(sessionRepo.findAll());
     }
 
@@ -162,7 +165,7 @@ public class SessionService {
     @Transactional
     public List<Session> findSessionsByInstructor(Instructor supervisor){
         if(supervisor == null){
-            throw new IllegalArgumentException();
+            throw new IllegalArgumentException("Instructor does not exist");
         }
         return sessionRepo.findBySupervisor(supervisor);
     }
@@ -184,10 +187,6 @@ public class SessionService {
 		}
 		return resultList;
 	}
-
-    public Course getCourseById(int id){
-        return courseRepo.findCourseById(id);
-    }
 
     public Instructor getInstructorById(int id){
         return instructRepo.findInstructorById(id);
